ISPN-11614 AbstractInternalProtoStreamMarshaller utilised by PersistenceMarshallerImpl
